# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CodeArena API",
    "version": "1.0.0",
    "description": "Real-time competitive programming platform API."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
  "tags": [
    {
      "name": "auth",
      "description": "Authentication and session management"
    },
    {
      "name": "match",
      "description": "Match lobby and lifecycle"
    },
    {
      "name": "health",
      "description": "Service health"
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "health",
        "summary": "Liveness and database connectivity check",
        "tags": [
          "health"
        ],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/github": {
      "get": {
        "operationId": "githubAuthRedirect",
        "summary": "Start GitHub OAuth flow",
        "tags": [
          "auth"
        ],
        "responses": {
          "302": {
            "description": "Redirect to GitHub authorization page"
          }
        }
      },
      "post": {
        "operationId": "githubCallback",
        "summary": "Complete GitHub OAuth flow",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GithubCallbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "Register a new account",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Log in with email and password",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "summary": "Rotate the refresh token cookie and issue a new access token",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Clear the refresh token and revoke the session",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "getMe",
        "summary": "Get the current authenticated user",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/matches": {
      "get": {
        "operationId": "listMatches",
        "summary": "List matches waiting for an opponent",
        "tags": [
          "match"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Open matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LobbyMatchList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "operationId": "createMatch",
        "summary": "Create a match and wait for an opponent",
        "tags": [
          "match"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Match created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LobbyMatch"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/matches/{id}/join": {
      "post": {
        "operationId": "joinMatch",
        "summary": "Join an open match",
        "tags": [
          "match"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][0-9a-z]{6,}$",
              "example": "clx0a1b2c3d4e5f6g7h8i9j0"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Joined"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/matches/{id}/cancel": {
      "post": {
        "operationId": "cancelMatch",
        "summary": "Cancel a match you are hosting",
        "tags": [
          "match"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][0-9a-z]{6,}$",
              "example": "clx0a1b2c3d4e5f6g7h8i9j0"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Cancelled"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GithubCallbackRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "example": "a1b2c3d4e5f6g7h8"
          },
          "state": {
            "type": "string",
            "minLength": 1,
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          }
        },
        "required": [
          "code",
          "state"
        ],
        "description": "GitHub OAuth callback payload"
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "example": "ada@example.com"
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "example": "ada"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72,
            "example": "supersecret"
          }
        },
        "required": [
          "email",
          "username",
          "password"
        ],
        "description": "New account registration payload"
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "example": "ada@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "example": "supersecret"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "description": "Email/password login payload"
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation error"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": "too_small"
                },
                "path": {
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      }
                    ]
                  },
                  "example": [
                    "password"
                  ]
                },
                "message": {
                  "type": "string",
                  "example": "Too small: expected string to have >=8 characters"
                }
              },
              "required": [
                "code",
                "path",
                "message"
              ],
              "additionalProperties": {}
            }
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "description": "Validation error with field-level issues"
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid email or password"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "description": "Standard error response"
      },
      "Health": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "db": {
            "type": "string",
            "const": "connected"
          }
        },
        "required": [
          "ok",
          "db"
        ],
        "additionalProperties": false,
        "description": "Service health status"
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ1c2VyIn0.sig"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "accessToken",
          "user"
        ],
        "additionalProperties": false,
        "description": "Access token with the authenticated user"
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx0a1b2c3d4e5f6g7h8i9j0"
          },
          "username": {
            "type": "string",
            "example": "ada"
          },
          "email": {
            "type": "string",
            "example": "ada@example.com"
          }
        },
        "required": [
          "id",
          "username",
          "email"
        ],
        "additionalProperties": false,
        "description": "Authenticated user"
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Logged out successfully"
          }
        },
        "required": [
          "message"
        ],
        "additionalProperties": false,
        "description": "Simple message acknowledgement"
      },
      "LobbyMatchList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/LobbyMatch"
        },
        "description": "Open matches, newest first"
      },
      "LobbyMatch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx0a1b2c3d4e5f6g7h8i9j0"
          },
          "host": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "clx9z8y7x6w5v4u3t2s1r0q"
              },
              "username": {
                "type": "string",
                "example": "ada"
              }
            },
            "required": [
              "id",
              "username"
            ],
            "additionalProperties": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "example": "2026-08-29T12:34:56.000Z"
          }
        },
        "required": [
          "id",
          "host",
          "createdAt"
        ],
        "additionalProperties": false,
        "description": "An open match waiting for an opponent"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request payload",
        "content": {
          "application/json": {
            "example": {
              "error": "Validation error"
            },
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or failed",
        "content": {
          "application/json": {
            "example": {
              "error": "Invalid email or password"
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "example": {
              "error": "Insufficient permissions"
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "example": {
              "error": "User not found"
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "Resource already exists",
        "content": {
          "application/json": {
            "example": {
              "error": "User with this email or username already exists"
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "example": {
              "error": "Internal Server Error"
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}